CHANGE: consolidate classes
