--- Assets/Scripts/Player/Health.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 59d7ab0102f548308bf9c9b3a4ea3026
-timeCreated: 1587073863
